api/requirements.txt
# For the web server framework
Flask

# For handling Cross-Origin Resource Sharing
Flask-Cors

# For numerical operations and image processing
numpy
Pillow